+++ exonum-java-binding/core/rust/src/proto/mod.rs
@@ -17,7 +17,6 @@
 #![allow(bare_trait_objects)]
 #![allow(renamed_and_removed_lints)]
 
-pub use self::service_runtime::*;
 pub use self::testkit::*;
 pub use exonum::proto::schema::*;