+++ exonum-java-binding/core/rust/src/storage/mod.rs
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 mod db;
-mod entry;
 mod fork;
 mod key_set_index;
 mod list_index;
 mod map_index;
 mod pair_iter;
+mod proof_entry;
 mod proof_list_index;
 mod proof_map_index;
 mod proof_map_index_next;
@@ -28,11 +28,11 @@ mod value_set_index;
 
 pub use self::db::Java_com_exonum_binding_core_storage_database_Views_nativeFree;
 pub(crate) use self::db::View;
-pub use self::entry::*;
 pub use self::key_set_index::*;
 pub use self::list_index::*;
 pub use self::map_index::*;
 pub use self::pair_iter::PairIter;
+pub use self::proof_entry::*;
 pub use self::proof_list_index::*;
 pub use self::proof_map_index::*;
 pub use self::proof_map_index_next::*;